+++ libraries/MCP23S17/MCP23S17.h
@@ -2,7 +2,7 @@
 //
 //    FILE: MCP23S17.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.5.1
+// VERSION: 0.5.2
 // PURPOSE: Arduino library for SPI MCP23S17 16 channel port expander
 //    DATE: 2021-12-30
 //     URL: https://github.com/RobTillaart/MCP23S17
@@ -10,10 +10,10 @@
 
 #include "Arduino.h"
 #include "SPI.h"
-#include "MCP23S17_registers.h"
+#include "MCP23x17_registers.h"
 
 
-#define MCP23S17_LIB_VERSION              (F("0.5.1"))
+#define MCP23S17_LIB_VERSION              (F("0.5.2"))
 
 //  ERROR CODES
 #define MCP23S17_OK                       0x00
@@ -24,8 +24,12 @@
 #define MCP23S17_REGISTER_ERROR           0xFF
 #define MCP23S17_INVALID_READ             0xFF
 
+
 #ifndef __SPI_CLASS__
-  #if defined(ARDUINO_ARCH_RP2040)
+  //  MBED must be tested before RP2040
+  #if defined(ARDUINO_ARCH_MBED)
+  #define __SPI_CLASS__   SPIClass
+  #elif defined(ARDUINO_ARCH_RP2040)
   #define __SPI_CLASS__   SPIClassRP2040
   #else
   #define __SPI_CLASS__   SPIClass
@@ -40,9 +44,9 @@ const uint32_t MCP23S17_MAX_SPI_SPEED = 10000000;
 class MCP23S17
 {
 public:
-  //  SOFTWARE SPI
+  //       SOFTWARE SPI
   MCP23S17(uint8_t select, uint8_t dataIn, uint8_t dataOut, uint8_t clock, uint8_t address = 0x00);
-  //  HARDWARE SPI
+  //       HARDWARE SPI
   MCP23S17(int select, __SPI_CLASS__ * spi);
   MCP23S17(int select, int address = 0x00, __SPI_CLASS__ * spi = &SPI);
 
@@ -51,8 +55,8 @@ public:
   uint8_t  getAddress();   //  default returns 0x00
 
 
-  //  single pin interface
-  //  mode: 0 = OUTPUT, 1 = INPUT, 1 = INPUT_PULLUP (==INPUT)
+  //       single pin interface
+  //       mode: 0 = OUTPUT, 1 = INPUT, 1 = INPUT_PULLUP (==INPUT)
   bool     pinMode1(uint8_t pin, uint8_t mode);
   bool     write1(uint8_t pin, uint8_t value);
   uint8_t  read1(uint8_t pin);
@@ -63,9 +67,9 @@ public:
   bool     getPullup(uint8_t pin, bool &pullup);
 
 
-  //  8 pins interface
-  //  port  = 0..1
-  //  value = bit pattern
+  //       8 pins interface
+  //       port  = 0..1
+  //       value = bit pattern
   bool     pinMode8(uint8_t port, uint8_t value);
   bool     write8(uint8_t port, uint8_t value);
   int      read8(uint8_t port);
@@ -76,8 +80,8 @@ public:
   bool     getPullup8(uint8_t port, uint8_t &mask);
 
 
-  //  16 pins interface
-  //  value = bit pattern
+  //       16 pins interface
+  //       value = bit pattern
   bool     pinMode16(uint16_t value);
   bool     write16(uint16_t value);
   uint16_t read16();
@@ -88,6 +92,29 @@ public:
   bool     getPullup16(uint16_t &mask);
 
 
+  //       INTERRUPTS (experimental)
+  //       pin = 0..15, mode = { RISING, FALLING, CHANGE }
+  bool     enableInterrupt(uint8_t pin, uint8_t mode);
+  bool     disableInterrupt(uint8_t pin);
+  
+  //       mask = 0x0000..0xFFFF  (overrides all earlier settings.
+  bool     enableInterrupt16(uint16_t mask, uint8_t mode);
+  bool     disableInterrupt16(uint16_t mask);
+
+  //       which pins caused the INT?
+  uint16_t getInterruptFlagRegister();
+  uint16_t getInterruptCaptureRegister();
+
+  //       polarity: 0 = LOW, 1 = HIGH, 2 = NONE/ODR
+  bool     setInterruptPolarity(uint8_t polarity);
+  uint8_t  getInterruptPolarity();
+
+  //       merge INTA and INTB
+  bool     mirrorInterrupts(bool on);
+  bool     isMirroredInterrupts();
+
+
+  //       SPI
   //       speed in Hz
   void     setSPIspeed(uint32_t speed);
   uint32_t getSPIspeed() { return _SPIspeed; };
@@ -96,22 +123,21 @@ public:
   bool     usesHWSPI() { return _hwSPI; };
   int      lastError();
 
-  //       set/clear IOCR bit fields  (0.2.3 experimental)
-  void     enableControlRegister(uint8_t mask);
-  void     disableControlRegister(uint8_t mask);
+  //       set/clear IOCR bit fields
+  bool     enableControlRegister(uint8_t mask);
+  bool     disableControlRegister(uint8_t mask);
   //       0.2.5 experimental
-  void     enableHardwareAddress();
-  void     disableHardwareAddress();
+  bool     enableHardwareAddress();
+  bool     disableHardwareAddress();
 
 
-private:
-  //       access to low level registers (just make these two functions public).
+protected:
+  //       access to low level registers (just make these functions public).
   //       USE WITH CARE !!!
   bool     writeReg(uint8_t reg, uint8_t value);
   uint8_t  readReg(uint8_t reg);
   bool     writeReg16(uint8_t reg, uint16_t value);
   uint16_t readReg16(uint8_t reg);
-
 
   uint8_t  _address = 0;
   uint8_t  _select  = 0;
@@ -122,7 +148,7 @@ private:
 
   bool     _hwSPI = true;
 
-  //  10 MHz is maximum, 8 is a better clock divider on AVR.
+  //       10 MHz is maximum, 8 is a better clock divider on AVR.
   uint32_t    _SPIspeed = MCP23S17_TYP_SPI_SPEED;
   __SPI_CLASS__ * _mySPI;
   SPISettings     _spi_settings;