+++ libraries/MCP23S17/CHANGELOG.md
@@ -6,11 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.2] - 2024-05-19
+- Fix #40, add several interrupt functions (experimental)
+- update **MCP23S17_registers.h** (reuse with MCP23017)
+- change return type of several functions
+  - e.g **bool enable/disableControlRegister()**
+- fix support for ARDUINO_ARCH_MBED
+- update readme.md
+- update keywords.txt
+
 ## [0.5.1] - 2024-03-02
 - Fix #38, add parameter to **begin(bool pullup)**
 - update GitHub/actions to version v4 in workflows.
 - Fix #37, using ints as parameter in constructor.
-
 
 ## [0.5.0] - 2024-01-20
 - Fix #32, improve handling SPI dependency.